zsmalloc: keep first object offset in struct page
In early draft of zspage migration, we couldn't use page._mapcount because
it was used for storing movable flag so we added runtime calculation to
get first object offset in a page but it causes rather many instruction
and even bug.

Since then, we don't use page._mapcount as page flag any more so now
there is no problem to use the field to store first object offset.

Link: http://lkml.kernel.org/r/1465788015-23195-1-git-send-email-minchan@kernel.org
Signed-off-by: Minchan Kim <minchan@kernel.org>
Cc: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>