mm, oom: hide mm which is shared with kthread or global init
The only case where the oom_reaper is not triggered for the oom victim is
when it shares the memory with a kernel thread (aka use_mm) or with the
global init.  After "mm, oom: skip vforked tasks from being selected" the
victim cannot be a vforked task of the global init so we are left with
clone(CLONE_VM) (without CLONE_SIGHAND).  use_mm() users are quite rare as
well.

In order to guarantee a forward progress for the OOM killer make sure that
this really rare cases will not get into the way and hide the mm from the
oom killer by setting MMF_OOM_REAPED flag for it.  oom_scan_process_thread
will ignore any TIF_MEMDIE task if it has MMF_OOM_REAPED flag set to catch
these oom victims.

After this patch we should guarantee a forward progress for the OOM killer
even when the selected victim is sharing memory with a kernel thread or
global init.

Link: http://lkml.kernel.org/r/1466426628-15074-11-git-send-email-mhocko@kernel.org
Signed-off-by: Michal Hocko <mhocko@suse.com>
Acked-by: Oleg Nesterov <oleg@redhat.com>
Cc: Vladimir Davydov <vdavydov@virtuozzo.com>
Cc: David Rientjes <rientjes@google.com>
Cc: Tetsuo Handa <penguin-kernel@I-love.SAKURA.ne.jp>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>