mm, page_alloc: don't retry initial attempt in slowpath
After __alloc_pages_slowpath() sets up new alloc_flags and wakes up
kswapd, it first tries get_page_from_freelist() with the new alloc_flags,
as it may succeed e.g.  due to using min watermark instead of low
watermark.  It makes sense to to do this attempt before adjusting zonelist
based on alloc_flags/gfp_mask, as it's still relatively a fast path if we
just wake up kswapd and successfully allocate.

This patch therefore moves the initial attempt above the retry label and
reorganizes a bit the part below the retry label.  We still have to
attempt get_page_from_freelist() on each retry, as some allocations cannot
do that as part of direct reclaim or compaction, and yet are not allowed
to fail (even though they do a WARN_ON_ONCE() and thus should not exist).
We can reuse the call meant for ALLOC_NO_WATERMARKS attempt and just set
alloc_flags to ALLOC_NO_WATERMARKS if the context allows it.  As a
side-effect, the attempts from direct reclaim/compaction will also no
longer obey watermarks once this is set, but there's little harm in that.

Kswapd wakeups are also done on each retry to be safe from potential races
resulting in kswapd going to sleep while a process (that may not be able
to reclaim by itself) is still looping.

Link: http://lkml.kernel.org/r/20160721073614.24395-4-vbabka@suse.cz
Signed-off-by: Vlastimil Babka <vbabka@suse.cz>
Acked-by: Michal Hocko <mhocko@suse.com>
Acked-by: Mel Gorman <mgorman@techsingularity.net>
Acked-by: David Rientjes <rientjes@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>