fcntl: add new F_OFD_*32 constants and handle them appropriately
When I originally added OFD lock support, we made the assumption that
userland would always pass in a struct flock64. It's possible however
for someone to build a binary without large file support (aka LFS),
which will call down into the kernel with the standard F_OFD_*
constants, but pass in a "legacy" struct flock.

My first thought was to patch glibc to cause a build break if anyone
tries to use OFD locks without LFS, but now I think it might be less
problematic to simply support OFD locks without LFS enabled.

The kernel handles this for classic POSIX locks with a separate set of
constants (postfixed with "64") to indicate that the incoming structure
is an LFS one. We can't do that here since the kernel already assumes
that the structure is an LFS one.

Instead, we can define a new set of constants that are postfixed with
"32" to indicate that the incoming structure is a non-LFS one. This
patch adds the kernel plumbing to handle that case. We'll also need a
small patch for glibc to make it to use these constants when LFS is
disabled.

In the event that someone builds a program with a new glibc, and runs it
on a kernel without support for F_OFD_*32 constants, they will just get
back EINVAL, which is preferable to the current situation (which is
undefined behavior due to misinterpretation of the struct flock
argument).

Cc: stable@vger.kernel.org	# v3.15+
Signed-off-by: Jeff Layton <jlayton@redhat.com>