fuse: update attributes before splice read
Before normal cached read fuse already updated attributes from userspace in
two cases:

 1) the end of the read was beyond current i_size

 2) in auto invalidate mode, to make sure we don't read stale data from the
    cache

The same logic would apply to reads with splice(2) as well, but in that
case the attribute refresh was missing.

This patch creates a helper, fuse_update_attr_before_read(), and calls this
in both the normal and the splice cases.

Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Fixes: a8894274a358 ("fuse: update attributes on aio_read")
Cc: <stable@vger.kernel.org> # v3.6+