tile: remove #pragma unroll from finv_buffer_remote()
This directive was put in the kernel source before the "pragma
unroll" support for tilegx gcc was upstreamed.  Remove it for
now, and we can put it back later if/when the compiler support
is upstreamed.  This avoids a warning when building the kernel.

This routine is not on a hot path in any case, so the extra
optimization here was mostly just for it's own sake.

Signed-off-by: Chris Metcalf <cmetcalf@mellanox.com>