leds: Introduce userspace LED class driver
This driver creates a userspace LED class driver similar to uinput.

New LED class devices are created by opening /dev/uleds and writing
a uleds_user_dev struct. A new LED class device is registered with the
name given in the struct. Reading will return a single byte that is
the current brightness. The poll() syscall is also supported. It will
be triggered whenever the brightness changes. Closing the file handle
to /dev/uleds will remove the LED class device.

Signed-off-by: David Lechner <david@lechnology.com>
Signed-off-by: Jacek Anaszewski <j.anaszewski@samsung.com>