UPSTREAM: clocksource/drivers/arm_arch_timer: Don't assume clock runs in suspend
The ARM specifies that the system counter "must be implemented in an
always-on power domain," and so we try to use the counter as a source of
timekeeping across suspend/resume. Unfortunately, some SoCs (e.g.,
Rockchip's RK3399) do not keep the counter ticking properly when
switched from their high-power clock to the lower-power clock used in
system suspend. Support this quirk by adding a new device tree property.

BUG=chrome-os-partner:57167
TEST=check RTC and system time before/after suspend, and see that system
     clock hasn't lagged (i.e., wasn't stopped in suspend):
     `date '+%s'; cat /sys/class/rtc/rtc0/since_epoch`

Change-Id: If9558bd6f8cc163d9a3b3b615354bc2eda5100d5
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Acked-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit d8ec7595a013)