UPSTREAM: platform/x86: intel_pmc_core: Fix PWRMBASE mask and mmio reg len
On Sunrise Point PCH, the Power Management Controller provides 4K bytes of
memory space for various power management and debug registers. This fix is
needed to access power management & debug registers that are mapped at a
higher offset.

Also, this provides a fix for correctly masking the PWRMBASE as the initial
bits (0-11) are reserved.

BUG=chrome-os-partner:62635
TEST=Intel PMC Core driver will be successfully loaded. Check for
TEST=echo freeze > /sys/power/state; wake the system and check for:
TEST=Above sysfs will provide the S0ix count which will be incremented

Change-Id: I353f1dbb9754ac23534d64ac2f3f7cd282977fdf
Signed-off-by: Rajneesh Bhardwaj <rajneesh.bhardwaj@intel.com>
Signed-off-by: Darren Hart <dvhart@linux.intel.com>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit 8434709ba714)