dm raid1: "mirror" target read error on default leg failure
In case the default leg of a "mirror" target fails, any read will
cause a new, operational default leg to be selected and the read be
resubmitted to it. Nonetheless that io succeeds, an error will be
reported back because bio->bi_error was not cleared properly.

Fix by clearing bio->bi_error before resubmission.

Resolves: rhbz1382516

Signed-off-by: Heinz Mauelshagen <heinzm@redhat.com>
Signed-off-by: Mike Snitzer <snitzer@redhat.com>