UPSTREAM: dt-bindings: rockchip-thermal: fix the misleading description
"rockchip,hw-tshut-temp", "rockchip,hw-tshut-mode" and
"rockchip,hw-tshut-polarity" are not a required properties
actually as the code could also work by loading the default
settings there. So it is apprently misleading, although we
prefer to get these from DT. And it seems we miss the 'rockchip,grf'
here which should also be an optional property.

BUG=chromium:707244
TEST=Build and run

Change-Id: I72e0e080bfcc9e80425a9a9845ed821a618346d6
Acked-by: Rob Herring <robh@kernel.org>
Reviewed-by: Heiko Stuebner <heiko@sntech.de>
Signed-off-by: Shawn Lin <shawn.lin@rock-chips.com>
Signed-off-by: Eduardo Valentin <edubezval@gmail.com>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit ecaa015c7ab3)