powerpc: factor out relocation code in module_64.c
The kexec_file_load system call needs to relocate the purgatory, so factor
out the module relocation code so that it can be shared.

This patch's purpose is to move the ELF relocation logic from
apply_relocate_add to the new function elf64_apply_relocate_add_item with
as few changes as possible.  The following changes were needed:

elf64_apply_relocate_add_item takes a my_r2 argument because the kexec
code can't use the my_r2 function since it doesn't have a struct module to
pass to it.  For the same reason, it also takes an obj_name argument to
use in error messages.  It still takes a pointer to struct module
argument, but kexec code can just pass NULL because except for the TOC
symbol, the purgatory doesn't have undefined symbols so the module pointer
isn't used.

Apart from what is described in the paragraph above, the code has no
functional changes.

Link: http://lkml.kernel.org/r/1477017617-8540-6-git-send-email-bauerman@linux.vnet.ibm.com
Signed-off-by: Thiago Jung Bauermann <bauerman@linux.vnet.ibm.com>
Suggested-by: Michael Ellerman <mpe@ellerman.id.au>
Cc: "Eric W. Biederman" <ebiederm@xmission.com>
Cc: Andreas Steffen <andreas.steffen@strongswan.org>
Cc: Balbir Singh <bsingharora@gmail.com>
Cc: Dave Young <dyoung@redhat.com>
Cc: Josh Sklar <sklar@linux.vnet.ibm.com>
Cc: Mimi Zohar <zohar@linux.vnet.ibm.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>