bluetooth: hack up ldisc to use serdev
This hacks up the BT ldisc in place to work as a serdev driver. It will
need refactoring into common, ldisc, and serdev parts.

It is working under QEMU to the point the driver can bind to a serio
device via DT, register as a BT device, and start to init a Broadcom
device with the serial port passed thru to the guest serial port. The
Broadcom device only works to this extent on my laptop.

Still need to figure out how to plumb a few tty things like
tty_unthrottle.

Signed-off-by: Rob Herring <robh@kernel.org>