CHROMIUM: drm: Allow render node access to KMS getters
We need to query the KMS API to present an accurate view of overlay
formats in Android (see CL:405019).  With the exception of
SET_CLIENT_CAP, these ioctls are getters.

SET_CLIENT_CAP is allowable with render nodes since it only presents
a different version of the drm core to userspace when modified.

BUG=chromium:616275
TEST=none

Change-Id: Id26e1b2b37dfa72f6ce55b524e849141750787c6
Reviewed-on: https://chromium-review.googlesource.com/405758
Commit-Ready: Gurchetan Singh <gurchetansingh@chromium.org>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>