CHROMIUM: tpm: cr50: Increase IRQ timeout
Increase the IRQ timeout (using TIS TIMEOUT A) to prevent issues
with TPM not being ready in time and add an error message if it
does not interrupt.

Since we are relying on an IRQ having a long timeout is safer than
having a short one.

Split the timeout for the "no interrupt" case so it does not make
transactions take multiple seconds.

BUG=chrome-os-partner:59191
TEST=suspend/resume testing on eve with cr50

Change-Id: I45bfde8193e51b6f93ce65babda5be32cc379091
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/406011
Reviewed-by: Aaron Durbin <adurbin@chromium.org>